requirements.txt
python-dotenv>=1.0.0
httpx>=0.27.0
openai>=1.30.0
apscheduler>=3.10.4
pyyaml>=6.0.1
pydantic>=2.7.0
akshare>=1.14.0
pandas>=2.2.0
ta>=0.11.0
jinja2>=3.1.4
pytest>=8.2.0
pytest-asyncio>=0.23.0
